Add type definitions for use in TypeScript
Previously, brace did not work well in TypeScript because the
existing type definitions for the Ace editor from DefinitelyTyped
are written assuming a global "ace" variable, rather than
being modular and thus importable like so:

    import * as ace from "brace";

Additionally, the changes that brace makes (mostly the acequire
rename) are not reflected in those definitions.

This commit adds a script to obtain the current type definitions
from DefinitelyTyped and rewrite them in a similar way as the
Ace source itself is rewritten in brace. The resulting
type definitions are exposed to TypeScript through the "typings"
key in package.json (though since they live in `index.d.ts`, they
would usually be discovered automatically).